// src/app/models/patient.model.ts
export class Patient {
    firstname: string;
    middlename: string;
    surname: string;
    dob: string;
    tribe: string;
    nationality: string;
    referencenumber: string;
    employername: string;
    employerlocation: string;
    employeraddress: string;
    employeremail: string;
    employertelephone: string;
    patientphonenumber: string;
    patientaddress: string;
    patientoccupation: string;
    billType: string;
    InsuranceID: string;
    emailaddress: string;
}